+++ src/Data.cpp
@@ -2,5 +2,5 @@
 
 #include "Data.hpp"
 
-Data::Data(std::vector<float> p_parameters, int p_label)
+Data::Data(std::vector<double> p_parameters, int p_label)
     : parameters(p_parameters), label(p_label) {}
